Fix code review comments
