+++ src/test/java/org/opensearch/knn/integ/PainlessScriptScoreIT.java
@@ -6,29 +6,25 @@
 package org.opensearch.knn.integ;
 
 import lombok.SneakyThrows;
-import org.apache.hc.core5.http.io.entity.EntityUtils;
-import org.opensearch.client.Request;
-import org.opensearch.client.Response;
-import org.opensearch.client.ResponseException;
 import org.opensearch.common.settings.Settings;
-import org.opensearch.common.xcontent.XContentFactory;
-import org.opensearch.core.rest.RestStatus;
-import org.opensearch.core.xcontent.ToXContent;
-import org.opensearch.core.xcontent.XContentBuilder;
-import org.opensearch.index.query.MatchAllQueryBuilder;
-import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.knn.KNNRestTestCase;
 import org.opensearch.knn.KNNResult;
-import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.engine.KNNMethodContext;
 import org.opensearch.knn.index.engine.MethodComponentContext;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
+import org.apache.hc.core5.http.io.entity.EntityUtils;
+import org.opensearch.client.Request;
+import org.opensearch.client.Response;
+import org.opensearch.client.ResponseException;
+import org.opensearch.index.query.MatchAllQueryBuilder;
+import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.core.rest.RestStatus;
+import org.opensearch.knn.integ.PainlessScriptHelper.MappingProperty;
 import org.opensearch.script.Script;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -36,43 +32,15 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 
 import static org.opensearch.knn.common.KNNConstants.METHOD_HNSW;
+import static org.opensearch.knn.integ.PainlessScriptHelper.createMapping;
 
-public class PainlessScriptIT extends KNNRestTestCase {
+public class PainlessScriptScoreIT extends KNNRestTestCase {
 
     public static final int AGGREGATION_FIELD_NAME_MIN_LENGTH = 2;
     public static final int AGGREGATION_FIELD_NAME_MAX_LENGTH = 5;
     private static final String NUMERIC_INDEX_FIELD_NAME = "price";
-
-    /**
-     * Utility to create a Index Mapping with multiple fields
-     */
-    protected String createMapping(List<MappingProperty> properties) throws IOException {
-        Objects.requireNonNull(properties);
-        XContentBuilder xContentBuilder = XContentFactory.jsonBuilder().startObject().startObject("properties");
-        for (MappingProperty property : properties) {
-            XContentBuilder builder = xContentBuilder.startObject(property.getName()).field("type", property.getType());
-            if (property.getDimension() != null) {
-                builder.field("dimension", property.getDimension());
-            }
-
-            if (property.getDocValues() != null) {
-                builder.field("doc_values", property.getDocValues());
-            }
-
-            if (property.getKnnMethodContext() != null) {
-                builder.startObject(KNNConstants.KNN_METHOD);
-                property.getKnnMethodContext().toXContent(builder, ToXContent.EMPTY_PARAMS);
-                builder.endObject();
-            }
-
-            builder.endObject();
-        }
-        xContentBuilder.endObject().endObject();
-        return xContentBuilder.toString();
-    }
 
     /*
      creates KnnIndex based on properties, we add single non-knn vector documents to verify whether actions
@@ -161,62 +129,12 @@ public class PainlessScriptIT extends KNNRestTestCase {
         deleteKNNIndex(INDEX_NAME);
     }
 
-    public void testCosineSimilarityScriptFields() throws Exception {
-        String source = String.format("1 + cosineSimilarity([2.0f, -2.0f], doc['%s'])", FIELD_NAME);
-        String scriptFieldName = "similarity";
-        Request request = buildPainlessScriptFieldsRequest(source, 3, getCosineTestData(), scriptFieldName);
-        Response response = client().performRequest(request);
-        assertEquals(request.getEndpoint() + ": failed", RestStatus.OK, RestStatus.fromCode(response.getStatusLine().getStatusCode()));
-
-        List<KNNResult> results = parseSearchResponseScriptFields(EntityUtils.toString(response.getEntity()), scriptFieldName);
-        assertEquals(3, results.size());
-
-        String[] expectedDocIDs = { "0", "1", "2" };
-        for (int i = 0; i < results.size(); i++) {
-            assertEquals(expectedDocIDs[i], results.get(i).getDocId());
-        }
-        deleteKNNIndex(INDEX_NAME);
-    }
-
-    public void testScriptFieldsGetValueReturnsDocValues() throws Exception {
-        String source = String.format("doc['%s'].value[0]", FIELD_NAME);
-        String scriptFieldName = "doc_value_field";
-        Map<String, Float[]> testData = getKnnVectorTestData();
-        Request request = buildPainlessScriptFieldsRequest(source, testData.size(), testData, scriptFieldName);
-
-        Response response = client().performRequest(request);
-        assertEquals(request.getEndpoint() + ": failed", RestStatus.OK, RestStatus.fromCode(response.getStatusLine().getStatusCode()));
-
-        List<KNNResult> results = parseSearchResponseScriptFields(EntityUtils.toString(response.getEntity()), scriptFieldName);
-        assertEquals(testData.size(), results.size());
-
-        String[] expectedDocIDs = { "1", "2", "3", "4" };
-        for (int i = 0; i < results.size(); i++) {
-            assertEquals(expectedDocIDs[i], results.get(i).getDocId());
-        }
-        deleteKNNIndex(INDEX_NAME);
-    }
-
     private Request buildPainlessScoreScriptRequest(String source, int size, Map<String, Float[]> documents) throws Exception {
         buildTestIndex(documents);
         QueryBuilder qb = new MatchAllQueryBuilder();
         return constructScriptScoreContextSearchRequest(
             INDEX_NAME,
             qb,
-            Collections.emptyMap(),
-            Script.DEFAULT_SCRIPT_LANG,
-            source,
-            size,
-            Collections.emptyMap()
-        );
-    }
-
-    private Request buildPainlessScriptFieldsRequest(String source, int size, Map<String, Float[]> documents, String scriptFieldName)
-        throws Exception {
-        buildTestIndex(documents);
-        return constructScriptFieldsContextSearchRequest(
-            INDEX_NAME,
-            scriptFieldName,
             Collections.emptyMap(),
             Script.DEFAULT_SCRIPT_LANG,
             source,
@@ -719,56 +637,6 @@ public class PainlessScriptIT extends KNNRestTestCase {
             return client().performRequest(request);
         } finally {
             deleteKNNIndex(INDEX_NAME);
-        }
-    }
-
-    static class MappingProperty {
-
-        private final String name;
-        private final String type;
-        private String dimension;
-
-        private KNNMethodContext knnMethodContext;
-        private Boolean docValues;
-
-        MappingProperty(String name, String type) {
-            this.name = name;
-            this.type = type;
-        }
-
-        MappingProperty dimension(String dimension) {
-            this.dimension = dimension;
-            return this;
-        }
-
-        MappingProperty knnMethodContext(KNNMethodContext knnMethodContext) {
-            this.knnMethodContext = knnMethodContext;
-            return this;
-        }
-
-        MappingProperty docValues(boolean docValues) {
-            this.docValues = docValues;
-            return this;
-        }
-
-        KNNMethodContext getKnnMethodContext() {
-            return knnMethodContext;
-        }
-
-        String getDimension() {
-            return dimension;
-        }
-
-        String getName() {
-            return name;
-        }
-
-        String getType() {
-            return type;
-        }
-
-        Boolean getDocValues() {
-            return docValues;
         }
     }
 }
